Refactor UI and update barcode handling
- Improved structure and layout of `index.html`, including a new modal for adding products.
- Enhanced `index.js` with new data properties and methods for product management and ratings; removed loading state management.
- Updated barcode property types from `int` to `long` to accommodate larger values.

+++ ProductLib/Models/Rating.cs
@@ -13,7 +13,7 @@ namespace BarcodeRatingLib.Models
         private string? _user;
 
         public int Id { get; set; }
-        public int ProductBarcode { get; set; }
+        public long ProductBarcode { get; set; }
         public int Score
         {
             get => _score;
